refactor: move errors to folder, constants, inject package.json version (#105)
- Add src/cli/errors/ for CliConfigError, ConversionError
- Add src/cli/constants/ with CLI_VERSION injected from package.json
- Refactor services to use new structure
- Update README.md with CLI documentation

Closes #105
